+++ packages/app-store/office365video/lib/index.ts
