+++ packages/app-store/office365video/package.json
