+++ packages/app-store/office365video/utils.ts
